Renamed test_base.py to test_logger.py and added test_base.py to test deprecation warning.
